Extract removed-member collection from TransformWorkerProgram
Move the listed members out of TransformWorkerProgram without changing their bodies.
